+++ packages/core/src/domain/telemetry/telemetryEvent.types.ts
@@ -1,4 +1,3 @@
-/* eslint-disable */
 /**
  * DO NOT MODIFY IT BY HAND. Run `yarn json-schemas:sync` instead.
  */
@@ -537,10 +536,14 @@ export type TelemetryBrowserFeaturesUsage =
   | StartDurationVital
   | StopDurationVital
   | AddDurationVital
+  | StartAction
+  | StopAction
+  | StartResource
+  | StopResource
 /**
  * Schema of mobile specific features usage
  */
-export type TelemetryMobileFeaturesUsage = AddViewLoadingTime | TrackWebView
+export type TelemetryMobileFeaturesUsage = AddViewLoadingTime | TrackWebView | AndroidNetworkInstrumentation
 
 /**
  * Schema of common properties of Telemetry events
@@ -580,6 +583,7 @@ export interface CommonTelemetryProperties {
     | 'unity'
     | 'kotlin-multiplatform'
     | 'electron'
+    | 'rum-cpp'
   /**
    * The version of the SDK generating the telemetry event
    */
@@ -914,6 +918,34 @@ export interface AddDurationVital {
   feature: 'add-duration-vital'
   [k: string]: unknown
 }
+export interface StartAction {
+  /**
+   * startAction API
+   */
+  feature: 'start-action'
+  [k: string]: unknown
+}
+export interface StopAction {
+  /**
+   * stopAction API
+   */
+  feature: 'stop-action'
+  [k: string]: unknown
+}
+export interface StartResource {
+  /**
+   * startResource API
+   */
+  feature: 'start-resource'
+  [k: string]: unknown
+}
+export interface StopResource {
+  /**
+   * stopResource API
+   */
+  feature: 'stop-resource'
+  [k: string]: unknown
+}
 export interface AddViewLoadingTime {
   /**
    * addViewLoadingTime API
@@ -938,5 +970,16 @@ export interface TrackWebView {
    * trackWebView API
    */
   feature: 'trackWebView'
+  [k: string]: unknown
+}
+export interface AndroidNetworkInstrumentation {
+  /**
+   * Android network instrumentation
+   */
+  feature: 'androidNetworkInstrumentation'
+  /**
+   * The network instrumentation API used
+   */
+  type: 'CRONET' | 'OKHTTP'
   [k: string]: unknown
 }