+++ packages/rum-core/src/domain/contexts/urlContexts.ts
@@ -6,6 +6,7 @@ import {
   HookNames,
   DISCARDED,
   mockable,
+  buildUrl,
 } from '@datadog/browser-core'
 import type { LocationChange } from '../../browser/locationChangeObservable'
 import type { LifeCycle } from '../lifeCycle'
@@ -39,8 +40,9 @@ export function startUrlContexts(
 
   let previousViewUrl: string | undefined
 
-  lifeCycle.subscribe(LifeCycleEventType.BEFORE_VIEW_CREATED, ({ startClocks }) => {
-    const viewUrl = mockable(location).href
+  lifeCycle.subscribe(LifeCycleEventType.BEFORE_VIEW_CREATED, ({ startClocks, url }) => {
+    const locationHref = mockable(location).href
+    const viewUrl = url !== undefined ? buildUrl(url, locationHref).href : locationHref
     urlContextHistory.add(
       buildUrlContext({
         url: viewUrl,