+++ scripts/lib/checkBrowserSdkPackageJsonFiles.ts
@@ -1,6 +1,6 @@
-import type { PackageJsonFile } from '../release/check-release.ts'
 import { browserSdkVersion as releaseVersion } from './browserSdkVersion.ts'
 import { printLog } from './executionUtils.ts'
+import type { PackageJsonInfo } from './filesUtils.ts'
 import { findPackageJsonFiles } from './filesUtils.ts'
 
 export function checkPackageJsonFiles(): void {
@@ -17,30 +17,31 @@ export function checkPackageJsonFiles(): void {
     checkPackageDependencyVersions(packageJsonFile)
   }
 }
-function checkPackageJsonVersion(packageJsonFile: PackageJsonFile): void {
-  if (packageJsonFile.content?.private) {
+function checkPackageJsonVersion(packageJsonInfo: PackageJsonInfo): void {
+  if (packageJsonInfo.content.private) {
     // The developer extension is a private package, but it should still have a version
     if (
-      packageJsonFile.content.version &&
-      packageJsonFile.content.name !== '@datadog/browser-sdk-developer-extension'
+      packageJsonInfo.content.version &&
+      packageJsonInfo.content.name !== '@datadog/browser-sdk-developer-extension' &&
+      packageJsonInfo.relativePath !== 'package.json'
     ) {
-      throw new Error(`Private package ${packageJsonFile.relativePath} should not have a version`)
+      throw new Error(`Private package ${packageJsonInfo.relativePath} should not have a version`)
     }
-  } else if (packageJsonFile.content.version !== releaseVersion) {
+  } else if (packageJsonInfo.content.version !== releaseVersion) {
     throw new Error(
-      `Invalid version for ${packageJsonFile.relativePath}: expected ${releaseVersion}, got ${packageJsonFile.content.version}`
+      `Invalid version for ${packageJsonInfo.relativePath}: expected ${releaseVersion}, got ${packageJsonInfo.content.version}`
     )
   }
 }
-function checkPackageDependencyVersions(packageJsonFile: PackageJsonFile): void {
-  if (packageJsonFile.content.private) {
+function checkPackageDependencyVersions(packageJsonInfo: PackageJsonInfo): void {
+  if (packageJsonInfo.content.private) {
     return
   }
 
   for (const dependencies of [
-    packageJsonFile.content.dependencies,
-    packageJsonFile.content.devDependencies,
-    packageJsonFile.content.peerDependencies,
+    packageJsonInfo.content.dependencies,
+    packageJsonInfo.content.devDependencies,
+    packageJsonInfo.content.peerDependencies,
   ]) {
     if (!dependencies) {
       continue
@@ -49,7 +50,7 @@ function checkPackageDependencyVersions(packageJsonFile: PackageJsonFile): void 
     for (const [dependencyName, dependencyVersion] of Object.entries(dependencies)) {
       if (isBrowserSdkPackageName(dependencyName) && dependencyVersion !== releaseVersion) {
         throw new Error(
-          `Invalid dependency version for ${dependencyName} in ${packageJsonFile.relativePath}: expected ${releaseVersion}, got ${dependencyVersion}`
+          `Invalid dependency version for ${dependencyName} in ${packageJsonInfo.relativePath}: expected ${releaseVersion}, got ${dependencyVersion}`
         )
       }
     }