+++ scripts/lib/packagesDirectoryNames.ts
@@ -1,3 +1,5 @@
 import { readdirSync } from 'node:fs'
 
-export const packagesDirectoryNames: string[] = readdirSync('packages')
+export const packagesDirectoryNames: string[] = readdirSync('packages', { withFileTypes: true })
+  .filter((entry) => entry.isDirectory())
+  .map((entry) => entry.name)